requirements.txt
aiohttp
aiosignal
annotated-types==0.5.0
anthropic==0.2.10
antlr4-python3-runtime==4.9.3
anyio
argon2-cffi
argon2-cffi-bindings
asttokens
async-timeout
attrs
backcall
backoff==2.2.1
beautifulsoup4
bleach
brotlipy==0.7.0
certifi
cffi
chardet==5.1.0
charset-normalizer
chromadb==0.3.29
click
clickhouse-connect==0.6.6
cohere==3.10.0
colorama
coloredlogs==15.0.1
contourpy==1.1.0
cryptography
cycler==0.11.0
dataclasses-json
debugpy
decorator
defusedxml
dill==0.3.6
dnspython==2.3.0
duckdb==0.8.1
effdet==0.4.1
email-validator==2.0.0.post2
entrypoints
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing
fastapi==0.85.1
fastjsonschema
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist
fsspec==2023.6.0
google-search-results==2.4.2
graphlib-backport==1.0.3
greenlet
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna
importlib-metadata
importlib-resources
InstructorEmbedding==1.0.1
iopath==0.1.10
ipykernel
ipython
ipython-genutils
itsdangerous==2.1.2
jedi
Jinja2
joblib==1.3.1
jsonschema
jupyter-events
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab-pygments
kiwisolver==1.4.4
langchain
layoutparser==0.3.4
lxml
lz4==4.3.2
manifest-ml==0.0.1
Markdown==3.4.3
MarkupSafe
marshmallow
marshmallow-enum
matplotlib==3.7.2
matplotlib-inline
mistune==0.8.4
mkl-fft==1.3.6
mkl-random
mkl-service==2.4.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict
mypy-extensions==0.4.3
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.1
nlpcloud==1.1.43
nltk==3.8.1
notebook
notebook_shim
numexpr
numpy
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic
opencv-python==4.8.0.74
openlm==0.0.5
openpyxl==3.1.2
orjson==3.9.2
overrides==7.3.1
packaging
pandas==2.0.3
pandocfilters
parso
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pickleshare
Pillow==10.0.0
pkgutil_resolve_name
platformdirs
portalocker==2.7.0
posthog==3.0.1
prometheus-client
prompt-toolkit
protobuf==4.23.4
psutil
pulsar-client==3.2.0
pure-eval
pycocotools==2.0.6
pycparser
pydantic
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
Pygments
pyOpenSSL
pypandoc==1.11
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent
PySocks
pytesseract==0.3.10
python-dateutil
python-docx==0.8.11
python-dotenv
python-json-logger
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pywin32==305.1
pywinpty
PyYAML==6.0
pyzmq
redis==4.6.0
regex
requests
rfc3339-validator
rfc3986-validator
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash
sentence-transformers==2.2.2
sentencepiece==0.1.99
six
sniffio
soupsieve
SQLAlchemy
sqlitedict==2.1.0
stack-data
starlette==0.20.4
sympy==1.12
tabulate==0.9.0
tenacity
terminado
threadpoolctl==3.2.0
tiktoken
timm==0.9.2
tinycss2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado
tqdm
traitlets
transformers==4.30.2
typing-inspect
typing_extensions
tzdata==2023.3
ujson==5.8.0
unstructured==0.8.1
unstructured-inference==0.5.5
urllib3
uvicorn==0.22.0
Wand==0.6.11
watchfiles==0.19.0
wcwidth
webencodings==0.5.1
websocket-client
websockets==11.0.3
win-inet-pton
xlrd==2.0.1
XlsxWriter==3.1.2
yarl
zipp
zstandard==0.21.0